Functional test improvements after miner-fs graph split

See merge request GNOME/tracker!236

+++ utils/trackertestutils/helpers.py
@@ -38,6 +38,10 @@ from . import psutil_mini as psutil
 log = logging.getLogger(__name__)
 
 
+TRACKER_DBUS_PREFIX = 'org.freedesktop.Tracker3'
+TRACKER_MINER_FS_BUSNAME = 'org.freedesktop.Tracker3.Miner.Files'
+
+
 class AwaitException(RuntimeError):
     pass
 
@@ -82,8 +86,8 @@ class await_insert():
 
     Use like this:
 
-        expected = 'a nfo:Document; nie:url <test://url>'
-        with self.tracker.await_update(expected) as resource:
+        expected = 'a nfo:Document; nie:isStoredAs <test://url>'
+        with self.tracker.await_update(DOCUMENTS_GRAPH, expected) as resource:
             # Create or update a file that's indexed by tracker-miner-fs.
             #
             # The context manager will not exit from the 'with' block until the
@@ -96,9 +100,10 @@ class await_insert():
     ensure_resource() if you just want to ensure that some data is present.
 
     """
-    def __init__(self, conn, predicates, timeout=DEFAULT_TIMEOUT,
+    def __init__(self, conn, graph, predicates, timeout=DEFAULT_TIMEOUT,
                  _check_inserted=True):
         self.conn = conn
+        self.graph = graph
         self.predicates = predicates
         self.timeout = timeout
         self._check_inserted = _check_inserted
@@ -113,8 +118,9 @@ class await_insert():
 
         if self._check_inserted:
             query_check = ' '.join([
-                'SELECT ?urn tracker:id(?urn) '
-                ' WHERE { '
+                'SELECT ?urn tracker:id(?urn) ',
+                f' FROM NAMED <{self.graph}> ',
+                ' WHERE { ',
                 '   ?urn a rdfs:Resource ; ',
                 self.predicates,
                 '}'
@@ -124,8 +130,9 @@ class await_insert():
                 raise AwaitException("Expected data is already present in the store.")
 
         query_filtered = ' '.join([
-            'SELECT ?urn tracker:id(?urn) '
-            ' WHERE { '
+            'SELECT ?urn tracker:id(?urn) ',
+            ' FROM NAMED <%s> ',
+            ' WHERE { ',
             '   ?urn a rdfs:Resource ; ',
             self.predicates,
             #'   FILTER (tracker:id(?urn) = ~id) '
@@ -140,11 +147,10 @@ class await_insert():
             for event in events:
                 if event.get_event_type() in [Tracker.NotifierEventType.CREATE,
                                               Tracker.NotifierEventType.UPDATE]:
-                    log.debug("Processing %s event for id %s", event.get_event_type(), event.get_id())
+                    log.debug("Processing %s event for id %s", event.get_event_type().value_nick, event.get_id())
                     #stmt.bind_int('~id', event.get_id())
                     #cursor = stmt.execute(None)
-                    stmt = query_filtered % event.get_id()
-                    log.debug("Running %s", stmt)
+                    stmt = query_filtered % (self.graph, event.get_id())
                     cursor = self.conn.query(stmt)
 
                     if cursor.next():
@@ -178,20 +184,21 @@ class await_insert():
         return True
 
 
-class await_update():
+class await_property_update():
     """Context manager to await data updates by Tracker miners & extractors.
 
     Use like this:
 
-        before = 'nie:url <test://url1>'
-        after = 'nie:url <test://url2>'
-        with self.tracker.await_update(resource_id, before, after):
+        before = 'nie:isStoredAs <test://url1>'
+        after = 'nie:isStoredAs <test://url2>'
+        with self.tracker.await_property_update(resource_id, before, after):
             # Trigger an update of the data.
 
     """
-    def __init__(self, conn, resource_id, before_predicates, after_predicates,
+    def __init__(self, conn, graph, resource_id, before_predicates, after_predicates,
                  timeout=DEFAULT_TIMEOUT):
         self.conn = conn
+        self.graph = graph
         self.resource_id = resource_id
         self.before_predicates = before_predicates
         self.after_predicates = after_predicates
@@ -205,45 +212,34 @@ class await_update():
         log.info("Awaiting update of resource id %s", self.resource_id)
 
         query_before = ' '.join([
-            'SELECT nie:isStoredAs(?urn) ?urn tracker:id(?urn) '
-            ' WHERE { '
+            'SELECT ?urn tracker:id(?urn) ',
+            ' FROM NAMED <%s> ',
+            ' WHERE { ',
             '   ?urn a rdfs:Resource ; ',
             self.before_predicates,
             '   . FILTER (tracker:id(?urn) = %s) '
             '}'
-        ]) % self.resource_id
+        ]) % (self.graph, self.resource_id)
+
         cursor = self.conn.query(query_before)
         if not cursor.next():
             raise AwaitException("Expected data is not present in the store.")
 
-        self.stored_as = cursor.get_string(0)[0]
-
-        query_on_create = 'SELECT nie:isStoredAs(tracker:uri(%s)) { }'
         query_after = ' '.join([
             'SELECT ?urn tracker:id(?urn) '
+            ' FROM NAMED <%s> ',
             ' WHERE { '
-            '   ?urn a rdfs:Resource ; ',
+           '   ?urn a rdfs:Resource ; ',
             self.after_predicates,
-            '   . FILTER (tracker:id(?urn) = %s) ',
+            '   . FILTER (tracker:id(?urn) = %s) '
             '}'
-        ])
+        ]) % (self.graph, self.resource_id)
 
         def match_cb(notifier, service, graph, events):
             for event in events:
-                log.debug("Processing %s event for id %s", event.get_event_type(), event.get_id())
-                if event.get_event_type() == Tracker.NotifierEventType.DELETE and event.get_id() == self.resource_id:
-                    self.resource_deleted = True
-                elif event.get_event_type() in [Tracker.NotifierEventType.CREATE,
-                                                Tracker.NotifierEventType.UPDATE]:
-                    if event.get_event_type() == Tracker.NotifierEventType.CREATE:
-                        if not self.resource_deleted:
-                            raise AwaitException("Received insert with no prior delete")
-                        cursor = self.conn.query(query_on_create % event.get_id())
-                        if cursor.next() and cursor.get_string(0)[0] == self.stored_as:
-                            self.resource_id = event.get_id()
-
-                    log.debug("Running %s", query_after % event.get_id())
-                    cursor = self.conn.query(query_after % event.get_id())
+                if event.get_event_type() == Tracker.NotifierEventType.UPDATE and event.get_id() == self.resource_id:
+                    log.debug("Processing %s event for id %s", event.get_event_type(), event.get_id())
+                    cursor = self.conn.query(query_after)
 
                     if cursor.next():
                         log.debug("Query matched!")
@@ -262,6 +258,105 @@ class await_update():
     def __exit__(self, etype, evalue, etraceback):
         if etype is not None:
             return False
+        while not self.matched:
+            self.loop.run_checked()
+
+        GLib.source_remove(self.timeout_id)
+        GObject.signal_handler_disconnect(self.notifier, self.signal_id)
+
+        return True
+
+
+class await_content_update():
+    """Context manager to await updates to file contents.
+
+    When a file is updated, the old information it contained is deleted from
+    the store, and the new information is inserted as a new resource.
+
+    """
+    def __init__(self, conn, graph, before_resource_id, before_predicates, after_predicates,
+                 timeout=DEFAULT_TIMEOUT):
+        self.conn = conn
+        self.graph = graph
+        self.before_resource_id = before_resource_id
+        self.before_predicates = before_predicates
+        self.after_predicates = after_predicates
+        self.timeout = timeout
+
+        self.loop = mainloop.MainLoop()
+        self.notifier = self.conn.create_notifier(Tracker.NotifierFlags.NONE)
+        self.matched = False
+
+        self.result = InsertedResource(None, 0)
+
+    def __enter__(self):
+        log.info("Awaiting delete of resource id %s and creation of a new one", self.before_resource_id)
+
+        query_before = ' '.join([
+            'SELECT nie:isStoredAs(?urn) ?urn tracker:id(?urn) '
+            ' FROM NAMED <%s> ',
+            ' WHERE { '
+            '   ?urn a rdfs:Resource ; ',
+            self.before_predicates,
+            '   . FILTER (tracker:id(?urn) = %s) '
+            '}'
+        ]) % (self.graph, self.before_resource_id)
+        cursor = self.conn.query(query_before)
+        if not cursor.next():
+            raise AwaitException("Expected data is not present in the store.")
+        file_url = cursor.get_string(0)[0]
+
+        query_after = ' '.join([
+            'SELECT ?urn tracker:id(?urn) '
+            ' FROM NAMED <%s> ',
+            ' WHERE { '
+            '   ?urn a rdfs:Resource ; ',
+            '      nie:isStoredAs <%s> ; ',
+            self.after_predicates,
+            '}'
+        ]) % (self.graph, file_url)
+
+        # When a file is updated, the DataObject representing the file gets
+        # an UPDATE notification. The InformationElement representing the
+        # content gets a DELETE and CREATE notification, because it is
+        # deleted and recreated. We detect the latter situation.
+
+        self.matched_delete = False
+        def match_cb(notifier, service, graph, events):
+            for event in events:
+                log.debug("Received %s event for id %s", event.get_event_type().value_nick, event.get_id())
+                if event.get_id() == self.before_resource_id and event.get_event_type() == Tracker.NotifierEventType.DELETE:
+                    log.debug("  Matched delete")
+                    self.matched_delete = True
+
+                # The after_predicates may match after the miner-fs creates
+                # the new resource, or they may only match once the extractor
+                # processes the resource. The latter will be an UPDATE event.
+                elif self.matched_delete and event.get_event_type() in [Tracker.NotifierEventType.CREATE, Tracker.NotifierEventType.UPDATE]:
+                    cursor = self.conn.query(query_after)
+
+                    if cursor.next():
+                        self.result.urn = cursor.get_string(0)[0]
+                        self.result.id = cursor.get_integer(1)
+                        log.debug("Query matched! Got new urn %s", self.result.urn)
+
+                        self.matched = True
+                        self.loop.quit()
+
+        def timeout_cb():
+            log.info("Timeout fired after %s seconds", self.timeout)
+            raise AwaitTimeoutException(
+                f"Timeout awaiting update of resource {self.before_resource_id} "
+                f"with URL {file_url} matching: {self.after_predicates}")
+
+        self.signal_id = self.notifier.connect('events', match_cb)
+        self.timeout_id = GLib.timeout_add_seconds(self.timeout, timeout_cb)
+
+        return self.result
+
+    def __exit__(self, etype, evalue, etraceback):
+        if etype is not None:
+            return False
 
         while not self.matched:
             self.loop.run_checked()
@@ -275,8 +370,9 @@ class await_update():
 class await_delete():
     """Context manager to await removal of a resource."""
 
-    def __init__(self, conn, resource_id, timeout=DEFAULT_TIMEOUT):
+    def __init__(self, conn, graph, resource_id, timeout=DEFAULT_TIMEOUT):
         self.conn = conn
+        self.graph = graph
         self.resource_id = resource_id
         self.timeout = timeout
 
@@ -288,13 +384,14 @@ class await_delete():
         log.info("Awaiting deletion of resource id %s", self.resource_id)
 
         query_check = ' '.join([
-            'SELECT ?urn tracker:id(?urn) '
-            ' WHERE { '
+            'SELECT ?urn tracker:id(?urn) ',
+            'FROM NAMED <%s> ',
+            ' WHERE { ',
             '   ?urn a rdfs:Resource ; ',
             '   . FILTER (tracker:id(?urn) = %s) '
             '}'
         ])
-        cursor = self.conn.query(query_check % self.resource_id)
+        cursor = self.conn.query(query_check % (self.graph, self.resource_id))
         if not cursor.next():
             raise AwaitException(
                 "Resource with id %i isn't present in the store.", self.resource_id)
@@ -344,38 +441,44 @@ class StoreHelper():
 
         self.conn = conn
 
-    def await_insert(self, predicates, timeout=DEFAULT_TIMEOUT):
+    def await_insert(self, graph, predicates, timeout=DEFAULT_TIMEOUT):
         """Context manager that blocks until a resource is inserted."""
-        return await_insert(self.conn, predicates, timeout)
+        return await_insert(self.conn, graph, predicates, timeout)
 
-    def await_update(self, resource_id, before_predicates, after_predicates,
+    def await_property_update(self, graph, resource_id, before_predicates, after_predicates,
                      timeout=DEFAULT_TIMEOUT):
-        """Context manager that blocks until a resource is updated."""
-        return await_update(self.conn, resource_id, before_predicates,
-                            after_predicates, timeout)
+        """Context manager that blocks until a resource property is updated."""
+        return await_property_update(self.conn, graph, resource_id, before_predicates,
+                                     after_predicates, timeout)
 
-    def await_delete(self, resource_id, timeout=DEFAULT_TIMEOUT):
+    def await_content_update(self, graph, before_resource_id, before_predicates, after_predicates,
+                     timeout=DEFAULT_TIMEOUT):
+        """Context manager that blocks until a resource is deleted and recreated."""
+        return await_content_update(self.conn, graph, before_resource_id, before_predicates,
+                                    after_predicates, timeout)
+
+    def await_delete(self, graph, resource_id, timeout=DEFAULT_TIMEOUT):
         """Context manager that blocks until a resource is deleted."""
-        return await_delete(self.conn, resource_id, timeout)
+        return await_delete(self.conn, graph, resource_id, timeout)
 
-    def ensure_resource(self, predicates, timeout=DEFAULT_TIMEOUT):
-        """Ensure that a resource matching 'predicates' exists.
+    def ensure_resource(self, graph, predicates, timeout=DEFAULT_TIMEOUT):
+        """Ensure that a resource matching 'predicates' exists in 'graph'.
 
         This function will block if the resource is not yet created.
 
         """
-        await_ctx_mgr = await_insert(self.conn, predicates, timeout, _check_inserted=False)
+        await_ctx_mgr = await_insert(self.conn, graph, predicates, timeout, _check_inserted=False)
         with await_ctx_mgr as resource:
             # Check if the data was committed *before* the function was called.
             query_initial = ' '.join([
                 'SELECT ?urn tracker:id(?urn) '
+                f' FROM NAMED <{graph}>',
                 ' WHERE { '
                 '   ?urn a rdfs:Resource ; ',
                 predicates,
                 '}'
             ])
 
-            log.debug("Running: %s", query_initial)
             cursor = self.conn.query(query_initial)
             if cursor.next():
                 resource.urn = cursor.get_string(0)[0]
@@ -422,13 +525,17 @@ class StoreHelper():
         else:
             raise Exception("Multiple entries for resource %s" % uri)
 
-    # FIXME: rename to get_resource_id_by_nepomuk_url !!
-    def get_resource_id(self, url):
+    def get_content_resource_id(self, url):
         """
-        Get the internal ID for a given resource, identified by URL.
+        Gets the internal ID for an nie:InformationElement resource.
+
+        The InformationElement represents data stored in a file, not
+        the file itself. The 'url' parameter is the URL of the file
+        that stores the given content.
+
         """
         result = self.query(
-            'SELECT tracker:id(?r) WHERE { ?r nie:isStoredAs/nie:url "%s" }' % url)
+            'SELECT tracker:id(?r) WHERE { ?r a nie:InformationElement; nie:isStoredAs "%s" }' % url)
         if len(result) == 1:
             return int(result[0][0])
         elif len(result) == 0:
@@ -489,9 +596,10 @@ class TrackerDBusSandbox:
 
         log.info("Looking for active Tracker processes on the bus")
         for busname in self.daemon.list_names_sync():
-            if busname.startswith('org.freedesktop.Tracker3'):
+            if busname.startswith(TRACKER_DBUS_PREFIX):
                 pid = self.daemon.get_connection_unix_process_id_sync(busname)
-                tracker_processes.append(pid)
+                if pid is not None:
+                    tracker_processes.append(pid)
 
         log.info("Terminating %i Tracker processes", len(tracker_processes))
         for pid in tracker_processes:
@@ -510,11 +618,21 @@ class TrackerDBusSandbox:
         log.info("Stopping D-Bus daemon for sandbox.")
         self.daemon.stop()
 
-    def stop_miner_fs(self):
-        log.info("Stopping tracker-miner-fs process.")
-        pid = self.daemon.get_connection_unix_process_id_sync('org.freedesktop.Tracker1.Miner.Files')
-        os.kill(pid, signal.SIGTERM)
-        psutil.wait_pid(pid)
+    def stop_daemon(self, busname):
+        """Stops the daemon that owns 'busname'.
+
+        This can be used if you want to force the miner-fs to exit, for
+        example.
+
+        """
+        log.info("Stopping daemon process that owns %s.", busname)
+        pid = self.daemon.get_connection_unix_process_id_sync(busname)
+        if pid:
+            os.kill(pid, signal.SIGTERM)
+            psutil.wait_pid(pid)
+            log.info("Process %i has stopped.", pid)
+        else:
+            log.info("Couldn't find a process owning %s.", busname)
 
     def get_connection(self):
         return self.daemon.get_connection()
